modificação de tipo

+++ aula17-06/aula17-06.py
@@ -1,8 +1,8 @@
 #lançamentos de notas 
-n1 = int(input("Digite a nota 1: "))
-n2 = int(input("Digite a nota 2: "))
-n3 = int(input("Digite a nota 3: "))
-n4 = int(input("Digite a nota 4: "))
+n1 = float(input("Digite a nota 1: "))
+n2 = float(input("Digite a nota 2: "))
+n3 = float(input("Digite a nota 3: "))
+n4 = float(input("Digite a nota 4: "))
 
 #calculos
 soma = n1 + n2 + n3 + n4
